Remove "engines" from package.json
Node will soon be releasing a new major version several times a year so the engines field is getting outdated quickly, and this project doesn't depend on many features of Node anyway. This commit is to make it so that downstream projects like react-native don't print a warning when people install it with io.js 3.0, Node.js 4.0, etc.

To help catch issues, the Travis config file now tests the latest version of io.js. It also contains `sudo: false` so that it runs on Travis' new docker infrastructure: http://docs.travis-ci.com/user/workers/container-based-infrastructure/